+++ doc/whats-new.md
@@ -2,11 +2,38 @@ What's new
 ==========
 
 
-0.3.2 (unreleased)
-------------------
+0.4.1 (26th May 2021)
+---------------------
+
+### Bug fixes
+
+- Check if an equilibrium can be gridded as a connected double-null before gridding.
+  Prevents creation of invalid grids where second X-point is outside the first
+  flux-surface in the SOL (#104)\
+  By [John Omotani](https://github.com/johnomotani)
+
+0.4.0 (26th May 2021)
+---------------------
+
+### Breaking changes
+
+- Method for calculating zShift made more accurate. This will change the calculated
+  zShift for the same input file (#101)
+- Default value for 'curvature_type' changed from "curl(b/B) with x-y derivatives" to
+  "curl(b/B)". New default should be more accurate and more consistent when changing
+  grid sizes, but will produce slightly different output from the same input file. If
+  the old behaviour is needed, set 'curvature_type = "curl(b/B) with x-y derivatives"'
+  explicitly (#100)
 
 ### New features
 
+- Accuracy of calculation of zShift improved by integrating on FineContours
+  rather than PsiContours (#101)\
+  By [John Omotani](https://github.com/johnomotani)
+- Circular, concentric flux surfaces magnetic geometry (#100)\
+  By [John Omotani](https://github.com/johnomotani)
+- Non-grid based method for calculating curvature (#100)\
+  By [John Omotani](https://github.com/johnomotani)
 - Parallelise most expensive loops (#99)\
   By [John Omotani](https://github.com/johnomotani)
 - Fix position of start and end points of contours when refining and add more tolerance